+++ OndselSolver/MBDynSystem.cpp
@@ -151,7 +151,7 @@ void MbD::MBDynSystem::outputFiles()
 	auto asmtMotions = asmtAsm->motions;
 	std::ofstream os(movFile);
 	os << std::setprecision(std::numeric_limits<double>::digits10 + 1);
-	for (int i = 1; i < asmtTimes->size(); i++)
+	for (size_t i = 1; i < asmtTimes->size(); i++)
 	{
 		for (auto& node : *nodes) {
 			node->outputLine(i, os);
@@ -241,7 +241,7 @@ void MbD::MBDynSystem::readElementsBlock(std::vector<std::string>& lines)
 
 void MbD::MBDynSystem::eraseComments(std::vector<std::string>& lines)
 {
-	for (int i = 0; i < lines.size(); i++)
+	for (size_t i = 0; i < lines.size(); i++)
 	{
 		auto line = lines[i];
 		auto it = line.find('#');
